SubProcessPosix: initial implementation
I can execute sub processes. However, I cannot read the output.

I think - that the reason is that on Posix, the read callback needs to
be set before we start the process (I think that by the time we set it,
the sub-process has ended).

Leaving this as a bug for now.